Event: added name and Description
Places: place now uses double instead of int
UserEvent: reworked representation of user signed to event in DB
WebApi: removed mock data and added db data

+++ src/EventPlanner.BL/DTO/PlaceDTO.cs
@@ -1,12 +1,11 @@
 ﻿
-
 namespace EventPlanner.BL.DTO
 {
     public class PlaceDTO
     {
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         public override bool Equals(object obj)
         {
